Extract sovereign specific multi-dimensional utils into separate file
