+++ crates/module-system/module-implementations/sov-evm/src/call.rs
@@ -14,7 +14,7 @@ use sov_address::{EthereumAddress, FromVmAddress};
 use sov_metrics::{save_elapsed, start_timer};
 use sov_modules_api::macros::{serialize, UniversalWallet};
 use sov_modules_api::prelude::UnwrapInfallible;
-use sov_modules_api::{Context, GasInfo, GasSpec, Spec, StateAccessor, TxState};
+use sov_modules_api::{Context, GasSpec, Spec, StateAccessor, TxState};
 #[cfg(feature = "native")]
 use std::convert::Infallible;
 
@@ -26,6 +26,7 @@ use crate::execution_config::EVM_EXECUTION_CONFIG;
 use crate::executor::{get_cfg_env, transact};
 #[cfg(feature = "native")]
 use crate::metrics::EvmTxMetrics;
+use crate::sov_fee_and_gas_utils::project_receipt_gas_from_actual_fee;
 use crate::{
     gas_metering_mode, BorshSpecId, ChainSpecUpdate, ContractCreationPolicy,
     ContractCreationPolicyUpdate, Evm, EvmChainSpec, EvmRuntimeConfig, EvmRuntimeConfigUpdate,
@@ -52,12 +53,6 @@ pub enum CallMessage<S: Spec> {
     Call(RlpEvmTransaction),
     /// Update the runtime configuration
     UpdateRuntimeConfig(EvmRuntimeConfigUpdate<S>),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct ProjectedReceiptGas {
-    gas_used: u64,
-    cumulative_gas_used: u64,
 }
 
 impl<S: Spec> Evm<S>
@@ -323,7 +318,7 @@ where
             .gas_info();
 
         if let Some(projected_gas) =
-            Self::project_receipt_gas_from_actual_fee(&pending_tx.receipt, &gas_info)?
+            project_receipt_gas_from_actual_fee::<S>(&pending_tx.receipt, &gas_info)?
         {
             pending_tx.receipt.gas_used = projected_gas.gas_used;
             pending_tx.receipt.receipt.cumulative_gas_used = projected_gas.cumulative_gas_used;
@@ -411,44 +406,6 @@ where
         sequencer_gas_used
             .checked_div(evm_gas_to_sequencer_gas_ratio)
             .expect("gas_to_charge_per_evm_gas() should not be zero")
-    }
-
-    fn project_receipt_gas_from_actual_fee(
-        receipt: &Receipt,
-        gas_info: &GasInfo<S::Gas>,
-    ) -> anyhow::Result<Option<ProjectedReceiptGas>> {
-        let tx_fee_paid = gas_info.gas_value;
-        if !crate::fee_activation::is_actual_fee_projection_height_active(receipt.block_number) {
-            return Ok(None);
-        }
-
-        if tx_fee_paid == sov_bank::Amount::ZERO {
-            return Ok(None);
-        }
-
-        let current_gas_used = receipt.gas_used;
-        // Intentional fail-closed behavior: once projection is active, we reject transactions
-        // whose charged gas dimensions cannot be reconciled into a single exact receipt gas value.
-        let projected_gas_used =
-            derive_receipt_gas_used_from_actual_fee(tx_fee_paid, gas_info.gas_price.as_ref())?;
-
-        if projected_gas_used == current_gas_used {
-            return Ok(None);
-        }
-
-        let previous_cumulative = receipt
-            .receipt
-            .cumulative_gas_used
-            .checked_sub(current_gas_used)
-            .context("EVM: receipt cumulative gas underflow while projecting exact fee")?;
-        let projected_cumulative = previous_cumulative
-            .checked_add(projected_gas_used)
-            .context("EVM: receipt cumulative gas overflow while projecting exact fee")?;
-
-        Ok(Some(ProjectedReceiptGas {
-            gas_used: projected_gas_used,
-            cumulative_gas_used: projected_cumulative,
-        }))
     }
 
     fn create_receipt(
@@ -557,47 +514,6 @@ where
 
         Ok(())
     }
-}
-
-fn derive_receipt_gas_used_from_actual_fee(
-    tx_fee_paid: sov_bank::Amount,
-    gas_price_per_dimension: &[sov_bank::Amount],
-) -> anyhow::Result<u64> {
-    let Some(uniform_gas_price) = gas_price_per_dimension.first() else {
-        bail!("EVM: gas price vector must have at least one dimension");
-    };
-
-    ensure!(
-        uniform_gas_price.0 > 0,
-        "EVM: cannot reconcile receipt from actual fee with zero gas price"
-    );
-    // Intentional: non-uniform per-dimension prices are treated as a hard correctness error.
-    // If we cannot derive one exact gas_used value, we reject instead of emitting mismatched fees.
-    ensure!(
-        gas_price_per_dimension
-            .iter()
-            .all(|price| price == uniform_gas_price),
-        "EVM: cannot reconcile receipt from actual fee with non-uniform gas prices: {gas_price_per_dimension:?}"
-    );
-
-    ensure!(
-        tx_fee_paid.0 % uniform_gas_price.0 == 0,
-        "EVM: tx fee {tx_fee_paid} is not divisible by uniform gas price {uniform_gas_price}"
-    );
-
-    let projected_gas_used_u128 = tx_fee_paid
-        .0
-        .checked_div(uniform_gas_price.0)
-        .expect("division by zero should be impossible because zero gas price is rejected above");
-    let projected_gas_used = u64::try_from(projected_gas_used_u128)
-        .context("EVM: projected receipt gas used does not fit in u64")?;
-
-    ensure!(
-        projected_gas_used > 0,
-        "EVM: projected receipt gas used cannot be zero for non-zero fee"
-    );
-
-    Ok(projected_gas_used)
 }
 
 pub(crate) fn verify_contract_creation_allowlist<
@@ -786,7 +702,7 @@ fn on_revert<S: Spec>(
     // Revert the sovereign SDK transaction only if
     // 1. We're in the sequencer
     // 2. The submitter of this transaction is the preferred sequencer
-    // 3. The preferred seuqencer is not configured to publish reverted transactions
+    // 3. The preferred sequencer is not configured to publish reverted transactions
     //
     // Reverting the tx *in the preferred sequencer* will cause it to be rejected and excluded from the batch. Reverting it in any other context
     // will simply cause it to be excluded from the EVM's record keeping.
@@ -824,38 +740,5 @@ mod tests {
         assert_eq!(get_spec_id(&spec, 1), SpecId::CONSTANTINOPLE);
         assert_eq!(get_spec_id(&spec, 2), SpecId::BERLIN);
         assert_eq!(get_spec_id(&spec, 3), SpecId::BERLIN);
-    }
-
-    #[test]
-    fn derive_receipt_gas_used_from_actual_fee_uses_uniform_price() {
-        let gas_used = derive_receipt_gas_used_from_actual_fee(
-            sov_bank::Amount::new(210_000),
-            &[sov_bank::Amount::new(10), sov_bank::Amount::new(10)],
-        )
-        .expect("uniform gas price should derive exact gas");
-
-        assert_eq!(gas_used, 21_000);
-    }
-
-    #[test]
-    fn derive_receipt_gas_used_from_actual_fee_rejects_non_uniform_prices() {
-        let err = derive_receipt_gas_used_from_actual_fee(
-            sov_bank::Amount::new(210_000),
-            &[sov_bank::Amount::new(10), sov_bank::Amount::new(11)],
-        )
-        .expect_err("non-uniform gas prices should fail hard");
-
-        assert!(err.to_string().contains("non-uniform gas prices"));
-    }
-
-    #[test]
-    fn derive_receipt_gas_used_from_actual_fee_rejects_zero_price() {
-        let err = derive_receipt_gas_used_from_actual_fee(
-            sov_bank::Amount::new(210_000),
-            &[sov_bank::Amount::ZERO, sov_bank::Amount::ZERO],
-        )
-        .expect_err("zero gas price should fail hard");
-
-        assert!(err.to_string().contains("zero gas price"));
     }
 }
